requirements.txt
streamlit
langchain-community
langchain-google-genai
langchain
python-dotenv
pypdf
PyPDF2
faiss-cpu